Created separate folder for all constants files. (#162)
