+++ src/components/AdminLoginForm.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import InputField from './form/InputField';
-import { formSubmitBtnText, goBackBtnText } from '../utils/yjsgConstants';
+import { formSubmitBtnText, goBackBtnText } from '../constants/yjsg';
 import Button from './common/Button';
