+++ src/components/SelectedStudentsActionWrapper.jsx
@@ -12,7 +12,7 @@ import UpdateIdCardStatusSelectedStudents from './UpdateIdCardStatusSelectedStud
 import { isBusCoordinatorsDataFailed } from '../reducers/assetFilesReducer';
 import Popup from './common/Popup';
 import Button from './common/Button';
-import { BUS_COORDINATOR_ERROR_MESSAGE } from '../utils/messagesConstants';
+import { BUS_COORDINATOR_ERROR_MESSAGE } from '../constants/messages';
 
 /**
  * SelectedStudentsActionWrapper render Export, Print Now, Print Later, Mark as Present and
