+++ src/components/StudentIdCardModal.jsx
@@ -5,7 +5,7 @@ import isEmpty from 'lodash/isEmpty';
 import upperFirst from 'lodash/upperFirst';
 import * as shortId from 'shortid';
 import Barcode from 'react-barcode';
-import { YJSG_ID_CARD_SMALL_HEADING, YJSG_ID_CARD_MAIN_HEADING } from '../utils/textConstants';
+import { YJSG_ID_CARD_SMALL_HEADING, YJSG_ID_CARD_MAIN_HEADING } from '../constants/text';
 
 import { getFormattedStudentId } from '../utils/dataGridUtils';
 import { getBusCoordinators } from '../reducers/assetFilesReducer';
