+++ src/components/routeComponents/SplashPagePrePopulated.jsx
@@ -37,7 +37,7 @@ import {
   newRegistrationBtnText,
   adminLoginBtnText,
   invalidAdminMsg, formSubmitBtnText,
-} from '../../utils/yjsgConstants';
+} from '../../constants/yjsg';
 import { setRegistrationData } from '../../utils/registrationFormUtils';
 
 // FixMe: This component is unnecessary.
