+++ src/constants/text.js
