+++ src/utils/dataGridUtils.js
@@ -2,7 +2,7 @@ import cloneDeep from 'lodash/cloneDeep';
 
 import {
   gridMetaData,
-} from '../components/GridData';
+} from '../constants/gridData';
 
 /**
  * manageStudentTableWidth method is called when we have to manage table width in grid page.
